Extract global style to component
Extract global style to component.
